feat: finish migrate commencer page
