+++ packages/app/src/app/(default)/login/page.tsx
@@ -2,10 +2,10 @@ import { fr } from "@codegouvfr/react-dsfr";
 import Alert from "@codegouvfr/react-dsfr/Alert";
 import { config } from "@common/config";
 import { type NextServerPageProps } from "@common/utils/next";
-import { CenteredContainer } from "@design-system";
+import { CenteredContainer, Heading } from "@design-system";
 import { notFound } from "next/navigation";
+import { getServerSession } from "next-auth";
 
-import { CheckLogged } from "./CheckLogged";
 import { EmailLogin } from "./EmailLogin";
 import { MonCompteProLogin } from "./MonCompteProLogin";
 
@@ -44,37 +44,43 @@ const LoginPage = async ({ searchParams }: NextServerPageProps<never, "callbackU
     notFound();
   }
 
+  const session = await getServerSession();
   const callbackUrl = typeof searchParams.callbackUrl === "string" ? searchParams.callbackUrl : "";
   const error = typeof searchParams.error === "string" ? searchParams.error : "";
 
   return (
     <CenteredContainer>
       <h1>{title}</h1>
-      <CheckLogged>
-        {error && (
-          <>
-            <Alert
-              title="Erreur de connexion"
-              description={signinErrors[error] || signinErrors.default}
-              severity="error"
-            />
-            <br />
-          </>
-        )}
+      {session?.user ? (
+        <Heading as="h4" text="Vous êtes déjà connecté·e." />
+      ) : (
+        <>
+          {error && (
+            <>
+              <Alert
+                title="Erreur de connexion"
+                description={signinErrors[error] || signinErrors.default}
+                severity="error"
+              />
+              <br />
+            </>
+          )}
 
-        <EmailLogin callbackUrl={callbackUrl} />
-        {config.env !== "prod" && (
-          <a href={mailDevUrl} target="_blank">
-            Debug: Ouvrir MailDev
-          </a>
-        )}
-        {config.ff.moncomptepro && (
-          <>
-            <hr className={fr.cx("fr-mt-4w")} />
-            <MonCompteProLogin callbackUrl={callbackUrl} />
-          </>
-        )}
-      </CheckLogged>
+          <EmailLogin callbackUrl={callbackUrl} />
+          {config.env !== "prod" && (
+            <a href={mailDevUrl} target="_blank">
+              Debug: Ouvrir MailDev
+            </a>
+          )}
+          {config.ff.moncomptepro && (
+            <>
+              <hr className={fr.cx("fr-mt-4w")} />
+              <MonCompteProLogin callbackUrl={callbackUrl} />
+              {/* TODO add "Qu'est-ce que MCP" */}
+            </>
+          )}
+        </>
+      )}
     </CenteredContainer>
   );
 };
